+++ code/com/jivesoftware/os/miru/miru-manage-deployable/src/main/java/com/jivesoftware/os/miru/manage/deployable/UnhealthyTopologyShiftPredicate.java
@@ -43,7 +43,7 @@ public class UnhealthyTopologyShiftPredicate implements ShiftPredicate {
             }
         }
 
-        float currentPct = (float) unhealthyCount / (float) numberOfReplicas;
+        float currentPct = unhealthyCount / numberOfReplicas;
         return (currentPct > unhealthyPct);
     }
 }
